@@ -1,32 +1,27 @@
 {
   "name": "frontend-tech-test",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/node": "^14.0.27",
-    "@types/react": "^16.9.44",
-    "@types/react-dom": "^16.9.8",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-helmet": "^6.1.0",
-    "react-scripts": "^4.0.3",
-    "styled-components": "^5.1.1",
-    "typescript": "^3.9.7"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "build:styles": "postcss src/styles/tailwind.css -o src/styles/style.css",
-    "prebuild": "npm run build:styles",
-    "prestart": "npm run build:styles",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -41,13 +36,17 @@
     ]
   },
   "devDependencies": {
-    "autoprefixer": "^10.1.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.5",
-    "postcss": "^8.2.0",
-    "postcss-cli": "^8.3.0",
-    "postcss-import": "^13.0.0",
-    "tailwind-scrollbar": "^1.0.1",
-    "tailwindcss": "^2.0.1"
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.35",
+    "@types/react-dom": "^18.0.11",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.21",
+    "postcss-cli": "^10.1.0",
+    "postcss-import": "^15.1.0",
+    "react-helmet": "^6.1.0",
+    "styled-components": "^5.3.9",
+    "tailwind-scrollbar": "^3.0.0",
+    "tailwindcss": "^3.3.1",
+    "typescript": "^4"
   }
 }
