@@ -1,7 +1,8 @@
 module.exports = {
-  plugins: [
-    require("postcss-import"),
-    require("tailwindcss"),
-    require("autoprefixer"),
-  ],
-};
+    plugins: [
+      require("postcss-import"),
+      require("tailwindcss"),
+      require("autoprefixer"),
+    ],
+  };
+  
