@@ -1,12 +1,12 @@
 import React from 'react';
-import { shallow } from '../../enzyme';
+import { render, screen } from '@testing-library/react';
 
 import TermsAndConditions from './index';
 
 describe('Terms And Conditions render test', () => {
     it('renders terms list', () => {
-        const wrapper = shallow(<TermsAndConditions />);
-        expect(wrapper.find('ul li.list-disc')).toBeDefined();
-        expect(wrapper.find('ul li')).toHaveLength(6);
+        render(<TermsAndConditions />);
+        const termsListItems = screen.getAllByRole('listitem');
+        expect(termsListItems).toHaveLength(6);
       });
 });
